requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.13.1
amqp==5.2.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
Babel==2.14.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
billiard==4.2.0
bio-embeddings @ git+https://github.com/sacdallago/bio_embeddings.git@efb9801f0de9b9d51d19b741088763a7d2d0c3a2
bio-embeddings-allennlp==0.9.2
bio-embeddings-bepler==0.0.1
bio-embeddings-CPCProt==0.0.1
bio-embeddings-deepblast==0.1.0
bio-embeddings-plus==0.1.1
bio-embeddings-tape-proteins==0.5
biopython==1.79
blis==0.7.11
boto3==1.34.25
botocore==1.34.25
catalogue==1.0.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
conllu==4.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1695534305529/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docutils==0.20.1
editdistance==0.6.2
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fair-esm==0.4.0
filelock==3.13.1
flaky==3.7.0
Flask==2.1.3
Flask-Cors==3.0.10
fonttools==4.47.2
freesasa==2.2.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gensim==4.3.2
gevent==23.9.1
greenlet==3.0.3
gunicorn==20.1.0
h5py==3.10.0
huggingface-hub==0.20.3
humanize==4.9.0
idna==3.6
imagesize==1.4.1
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==4.13.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1705417941265/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1694607144474/work
itsdangerous==2.1.2
jax==0.4.23
jax-unirep==2.2.0
jaxlib==0.4.23
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonnet==0.20.0
jsonpickle==3.0.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1704727023078/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1694598704522/work
kiwisolver==1.4.5
kombu==5.3.5
lightdock==0.9.4
lightning-utilities==0.10.1
llvmlite==0.41.1
lmdb==1.4.1
lock==2018.3.25.2110
Mako==1.3.1
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
ml-dtypes==0.3.2
multidict==6.0.4
multipledispatch==1.0.0
murmurhash==1.0.10
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
nltk==3.8.1
numba==0.58.1
numpy==1.23.5
numpydoc==1.6.0
opt-einsum==3.3.0
optuna==3.5.0
overrides==4.1.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==1.5.3
parsimonious==0.10.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
plac==1.1.3
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
plotly==5.18.0
pluggy==1.3.0
preshed==3.0.9
ProDy==2.4.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.25.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722404069/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyfaidx==0.6.4
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pynndescent==0.5.11
pyparsing==3.1.1
pytest==7.4.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-slugify==6.1.2
pytorch-lightning==1.8.6
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2023.3.post1
PyYAML==6.0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1701783166972/work
regex==2023.12.25
requests==2.31.0
responses==0.24.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
sentencepiece==0.1.96
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.4.0
snowballstemmer==2.2.0
spacy==2.3.9
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.25
sqlparse==0.4.4
srsly==1.0.7
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==7.4.6
threadpoolctl==3.2.0
tokenizers==0.15.1
tomli==2.0.1
torch==1.9.1
torchmetrics==1.2.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1695373443748/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
transformers==4.37.0
typing-utils==0.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1702176139754/work
tzdata==2023.4
umap-learn==0.5.5
Unidecode==1.3.8
urllib3==1.26.18
vapeplot==0.0.8
vespa-effect==1.0.2
vine==5.1.0
wasabi==0.10.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==2.0.3
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1694598693908/work
word2number==1.1
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
zope.event==5.0
zope.interface==6.1
